requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-phonenumber-field==7.2.0
django-rest-framework-nested==0.0.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.10.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.1
phonenumbers==8.13.22
Pillow==10.0.1
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.4
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
